Local IP detection fix
Fixes a problem also shared by BloomDesktop (for which PR #6811 is currently active): the wrong local IP is presented to the Android device. Returning the first address found of type 'AddressFamily.InterNetwork' that is also "Up" is sometimes not correct. On my machine (which has 11 IP addresses) it is never correct. New algorithm is documented in code comments.
This commit is prior to any code cleanup being done.