IP selection logic revised, debug version
The revisions from the last working session with Tom B are incorporated, as much as possible. The new algorithm is described in AndroidSynchronization.cs comments in lines 164-184.
Works in my home office on both wired and WiFi (which are on same subnet). Still need to test at JAARS where wired and WiFi are on different subnets.
Code is pretty clean but still has full debug output.